gen/requirements.txt
PyYAML==6.0.1
requests==2.25.1
Flask==2.2.2
pyopenssl==19.1.0
openpyxl==3.0.9
cryptography==3.4.8
Werkzeug==2.2.2
